Add progress indicator when using kernel exec API
